requirements.txt
requests>=2.28.0
reportlab>=3.6.0
arabic-reshaper>=2.1.3
python-bidi>=0.4.2
unidecode>=1.3.0 